requirements.txt
pdfplumber==0.10.3
requests==2.31.0
supabase==2.3.0
schedule==1.2.0
selenium==4.15.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0